Initial SASS conversion commit
SASS is a CSS extension that makes writing CSS easy, clean, and programmatic
through the use of things like variables and mixins. It's real easy to pick
up and learn - just check out sass-lang.com. HTML pages don't use SASS files
directly, though - they still need to read CSS. Thus, SASS is a preprocessor,
and needs to be compiled into CSS.

Unfortunately, it's sort of easier to start a project from scratch using
SASS than trying to convert a large existing CSS project into SASS and make
it all pretty. This first commit is simply what an automatic CSS-to-SASS
program spat out. The work of renaming variables and restructuring code to be
clean and legible still needs to be done by hand. This is probably a simple
thing for anyone to do when they're looking for some not-super-difficult
work to do. I guess that's what this branch is about.